current cloud-interface branch code failed to build iotivity stack.
so I merge latest master code that the build issue is resolved 
to build full source code.

Reviewed-on: https://gerrit.iotivity.org/gerrit/2265
Tested-by: Jon A. Cruz <jonc@osg.samsung.com>

+++ resource/csdk/security/src/pstatresource.c
@@ -254,7 +254,7 @@ static OCEntityHandlerResult HandlePstatPutRequest(const OCEntityHandlerRequest 
              */
             for(size_t i=0; i< gPstat->smLen; i++)
             {
-                if(gPstat->sm[i] == omJson->valueint)
+                if(gPstat->sm[i] == (unsigned int)omJson->valueint)
                 {
                     gPstat->om = (OicSecDpom_t)omJson->valueint;
                     break;
